#282 - Refactor: home과 home-empty 페이지 병합
