+++ src/App.jsx
@@ -20,7 +20,7 @@ import Follower from './pages/Follower/Follower.jsx';
 import Following from './pages/Following/Following.jsx';
 import FollowListPage from './pages/FollowListPage/FollowListPage.jsx';
 import AddProduct from './pages/AddProduct/AddProduct.jsx';
-import HomeEmpty from './pages/HomeEmpty/HomeEmpty';
+import HomeFeed from './pages/HomeFeed/HomeFeed';
 import Search from './pages/Search/Search';
 import NotFound from './pages/NotFound/NotFound';
 import Home from './pages/Home/Home';
@@ -36,7 +36,7 @@ function App() {
             <Route path="/login" exact component={Login} />
             <Route path="/join" exact component={Join} />
             <Route path="/join/profile" exact component={ProfileSetting} />
-            <Route path="/home-empty" exact component={HomeEmpty} />
+            <Route path="/home-feed" exact component={HomeFeed} />
             <Route path="/post/:postId" exact component={Post} />
             <Route path="/post/:postId/update" component={PostUpdate} exact />
             <Route path="/upload" exact component={Upload} />
